feat(core)!: shared receiver-side task store with spec fixes (#143 phase 2)
Move the receiver-side task machinery (TaskManager, TaskHandle, TaskState,
route_task_store, CancellationToken) from mcpkit-server to a new
mcpkit_core::tasks module so the client (phase 3, task-augmented sampling)
can reuse it. mcpkit-server re-exports everything at its existing paths.

The move fixes three spec deviations in the shared store, for the stdio
runtime and all four HTTP adapters at once:

- tasks/result now BLOCKS until the task reaches a terminal status (spec
  MUST; previously -32602 'task is not completed'). The wait parks on a
  per-task event-listener Event with no lock held across an await; the
  runtime's cooperative loop and concurrent requests stay live while a
  tasks/result waits, and tasks/cancel unblocks waiters.
- tasks/result reproduces the underlying request's outcome exactly. The
  terminal payload is now two-variant (TaskPayload::Success /
  TaskPayload::Error(JsonRpcError)); a failed request's JSON-RPC error is
  stored via TaskHandle::fail_with_error and returned verbatim (code,
  message, data) through the new pass-through McpError::JsonRpc variant.
  Per spec, a tools/call result with isError:true now moves the task to
  'failed' (TaskHandle::fail_with_result) while tasks/result still returns
  that result.
- Successful tasks/result responses carry
  _meta['io.modelcontextprotocol/related-task'] (spec MUST), merged with any
  existing _meta. Error responses carry none: the schema's Error object has
  no _meta slot, so the requirement is unsatisfiable there (documented).

BREAKING: TaskState.payload is Option<TaskPayload> (was Option<Value>),
TaskManager::payload returns the new type, and route_task_store is async.

Refs #143 (phase 2 of 3 - leave open).
