+++ crates/mcpkit-axum/src/handler.rs
@@ -290,7 +290,7 @@ where
                         // Not task-augmented — fall through to the normal path.
                         AugmentedTaskOutcome::NotApplicable => {}
                     }
-                } else if let Some(result) = route_task_store(store, method, params) {
+                } else if let Some(result) = route_task_store(store, method, params).await {
                     return match result {
                         Ok(value) => Response::success(request.id.clone(), value),
                         Err(e) => Response::error(request.id.clone(), e.into()),
